requirements.txt
fastapi>=0.115.9
uvicorn>=0.15.0
pydantic>=2.11.1
python-dotenv>=1.1.0

# Google services
google-generativeai>=0.8.5
google-api-python-client>=2.166.0
firebase-admin

# Database
sqlalchemy>=1.4.0

# Development tools
pytest>=6.0.0
black>=21.0.0
isort>=5.0.0
mypy>=0.910

# HTTP client
requests>=2.32.3

# Utilities
python-dateutil>=2.9.0

annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
azure-ai-ml==1.27.0
azure-ai-projects==1.0.0b10
azure-cognitiveservices-speech==1.43.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.22.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
azure-storage-file-share==12.21.0
azureml==0.2.7
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
Caffe==0.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
durationpy==0.9
executing==2.2.0
filelock==3.17.0
fixedint==0.1.6
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.9.0
streamlit>=1.28.0
requests>=2.31.0
websocket-client>=1.6.0

# Vector database and embeddings for conversation history
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0 